--- a/src/System.IO.Pipelines/FlushResult.cs
+++ b/src/System.IO.Pipelines/FlushResult.cs
@@ -5,20 +5,16 @@ namespace System.IO.Pipelines
 {
     public struct FlushResult
     {
-        public FlushResult(bool isCancelled, bool isCompleted)
-        {
-            IsCancelled = isCancelled;
-            IsCompleted = isCompleted;
-        }
+        internal ResultFlags ResultFlags;
 
         /// <summary>
         /// True if the currrent flush was cancelled
         /// </summary>
-        public bool IsCancelled { get; }
+        public bool IsCancelled => (ResultFlags & ResultFlags.Cancelled) > 0;
 
         /// <summary>
         /// True if the <see cref="IPipeWriter"/> is complete
         /// </summary>
-        public bool IsCompleted { get; }
+        public bool IsCompleted => (ResultFlags & ResultFlags.Completed) > 0;
     }
 }