--- a/src/System.IO.Pipelines/ReadResult.cs
+++ b/src/System.IO.Pipelines/ReadResult.cs
@@ -34,11 +34,11 @@ namespace System.IO.Pipelines
         /// <summary>
         /// True if the currrent read was cancelled
         /// </summary>
-        public bool IsCancelled => (ResultFlags & ResultFlags.Cancelled) > 0;
+        public bool IsCancelled => (ResultFlags & ResultFlags.Cancelled) != 0;
 
         /// <summary>
         /// True if the <see cref="IPipeReader"/> is complete
         /// </summary>
-        public bool IsCompleted => (ResultFlags & ResultFlags.Completed) > 0;
+        public bool IsCompleted => (ResultFlags & ResultFlags.Completed) != 0;
     }
 }