--- a/src/System.IO.Pipelines/ResultFlags.cs
+++ b/src/System.IO.Pipelines/ResultFlags.cs
@@ -1,0 +1,10 @@
+namespace System.IO.Pipelines
+{
+    [Flags]
+    internal enum ResultFlags : byte
+    {
+        None = 0,
+        Cancelled = 1,
+        Completed = 1 >> 2
+    }
+}