Before:

```
                  Method |        Mean |    StdErr |     StdDev |          RPS |
------------------------ |------------ |---------- |----------- |------------- |
 ParseLiveAspNetTwoTasks | 326.4556 ns | 3.3105 ns | 18.1321 ns | 3,063,203.49 |
   ParseLiveAspNetInline | 237.0526 ns | 3.8329 ns | 20.9937 ns | 4,218,472.53 |
```

After:
```
                 Method |        Mean |    StdErr |     StdDev |          RPS |
------------------------ |------------ |---------- |----------- |------------- |
 ParseLiveAspNetTwoTasks | 372.2635 ns | 3.6436 ns | 19.9568 ns | 2,686,269.24 |
   ParseLiveAspNetInline | 172.7343 ns | 3.5178 ns | 19.2678 ns | 5,789,239.57 |
```