@pakrym 2017-03-23
@davidfowl not sure, but it heavily relies on threadpool so I don't trust it as much